Move remainder of `api_common.php` to `common.php`
