+++ app/Http/Controllers/ViewTestController.php
@@ -9,8 +9,6 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\View\View;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 
-require_once 'include/api_common.php';
-
 final class ViewTestController extends AbstractBuildController
 {
     public function viewTest(): View
