+++ app/cdash/app/Controller/Api/TestOverview.php
@@ -22,7 +22,6 @@ use CDash\Model\BuildGroup;
 use CDash\Model\Project;
 
 require_once 'include/filterdataFunctions.php';
-require_once 'include/api_common.php';
 
 class TestOverview extends ResultsApi
 {
