+++ app/cdash/public/api/v1/index.php
@@ -17,7 +17,6 @@
 
 namespace CDash\Api\v1\Index;
 
-require_once 'include/api_common.php';
 require_once 'include/filterdataFunctions.php';
 
 use CDash\Controller\Api\Index as IndexController;
